TypeScript-ify Alert component
