+++ app/javascript/packages/components/alert.spec.tsx
@@ -1,6 +1,6 @@
 import { createRef } from 'react';
-import { Alert } from '@18f/identity-components';
-import { render } from '../../support/document-capture';
+import { render } from '@testing-library/react';
+import Alert from './alert';
 
 describe('identity-components/alert', () => {
   it('should apply alert role', () => {
